requirements.txt
flask==2.0.1
flask-cors==3.0.10
docker==6.1.3
werkzeug==2.0.1
python-dotenv==0.19.0
PyYAML==6.0.1